refactor: split topbar stylesheet to another file
